pyproject.toml
[project]
name = "shadowgen-ml-training"
version = "0.1.0"
description = "Training-side scaffold for ShadowGen shadow generation models."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0.0",
    "pillow>=11.0.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
shadowgen-training = "shadowgen_training.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/shadowgen_training"]
